1.修改部署文件 2.新增一些函数备注 3.修改administrated.sol

+++ contracts/ERC223RecevingContract.sol
@@ -13,5 +13,5 @@ contract ERC223ReceivingContract {
      * @param _value Amount of tokens.
      * @param _data  Transaction metadata.
      */
-    function tokenFallback(address _from, uint _value, bytes _data) public;
+    function tokenFallback(address _from, uint256 _value, bytes _data) public;
 }
